requirements.txt
anthropic[bedrock]>=0.40.0
boto3>=1.35.0
pypdf>=4.0.0
pdfplumber>=0.11.0
pypdfium2>=4.18.0
Pillow>=10.0.0
